move settings-panel under details-view-overlay
