+++ src/DetailsView/components/details-view-overlay/settings-panel/settings-panel.scss
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-@import '../../../common/styles/colors.scss';
+@import '../../../../common/styles/colors.scss';
 
 .settings-panel {
     h3 {
