+++ src/DetailsView/components/details-view-overlay/settings-panel/settings-panel.tsx
@@ -4,8 +4,8 @@ import { NamedFC } from 'common/react/named-fc';
 import { FeatureFlagStoreData } from 'common/types/store-data/feature-flag-store-data';
 import { UserConfigurationStoreData } from 'common/types/store-data/user-configuration-store';
 import * as React from 'react';
-import { DetailsViewActionMessageCreator } from '../../actions/details-view-action-message-creator';
-import { GenericPanel } from '../generic-panel';
+import { DetailsViewActionMessageCreator } from '../../../actions/details-view-action-message-creator';
+import { GenericPanel } from '../../generic-panel';
 import * as styles from './settings-panel.scss';
 import { SettingsDeps } from './settings/settings-props';
 import { SettingsComponent, SettingsProvider } from './settings/settings-provider';
