+++ src/DetailsView/components/details-view-overlay/settings-panel/settings/settings-provider-impl.ts
