+++ src/DetailsView/components/details-view-overlay/settings-panel/settings/settings-provider.tsx
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-import { ReactFCWithDisplayName } from '../../../../common/react/named-fc';
+import { ReactFCWithDisplayName } from '../../../../../common/react/named-fc';
 import { SettingsProps } from './settings-props';
 
 export type SettingsComponent = ReactFCWithDisplayName<SettingsProps>;
