+++ src/issue-filing/types/settings-form-props.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 import { SetIssueFilingServicePropertyPayload } from 'background/actions/action-payloads';
-import { SettingsDeps } from 'DetailsView/components/settings-panel/settings/settings-props';
+import { SettingsDeps } from 'DetailsView/components/details-view-overlay/settings-panel/settings/settings-props';
 
 export type SettingsFormProps<Settings> = {
     deps: SettingsDeps;
