+++ src/tests/unit/tests/DetailsView/components/settings-panel/settings-panel.test.tsx
@@ -4,9 +4,13 @@ import { UserConfigMessageCreator } from 'common/message-creators/user-config-me
 import { NamedFC, ReactFCWithDisplayName } from 'common/react/named-fc';
 import { UserConfigurationStoreData } from 'common/types/store-data/user-configuration-store';
 import { DetailsViewActionMessageCreator } from 'DetailsView/actions/details-view-action-message-creator';
-import { SettingsPanel, SettingsPanelDeps, SettingsPanelProps } from 'DetailsView/components/settings-panel/settings-panel';
-import { SettingsProps } from 'DetailsView/components/settings-panel/settings/settings-props';
-import { createSettingsProvider } from 'DetailsView/components/settings-panel/settings/settings-provider';
+import {
+    SettingsPanel,
+    SettingsPanelDeps,
+    SettingsPanelProps,
+} from 'DetailsView/components/details-view-overlay/settings-panel/settings-panel';
+import { SettingsProps } from 'DetailsView/components/details-view-overlay/settings-panel/settings/settings-props';
+import { createSettingsProvider } from 'DetailsView/components/details-view-overlay/settings-panel/settings/settings-provider';
 import { shallow } from 'enzyme';
 import * as React from 'react';
